requirements.txt
Django==1.9.5
django-allauth==0.25.2
django-appconf==1.0.2
django-bootstrap-form==3.2
django-bower==5.1.0
django-braces==1.8.1
django-haystack==2.4.1
django-model-utils==2.4
django-social-auth==0.7.28
djangorestframework==3.3.3
httplib2==0.9.2
jsonfield==1.0.3
oauth2==1.9.0.post1
oauthlib==1.0.3
Pillow==3.1.1
psycopg2==2.6.1
PyJWT==1.4.0
python-instagram==1.3.2
python-openid==2.2.5
python-social-auth==0.2.14
pytz==2016.3
requests==2.9.1
requests-oauthlib==0.6.1
simplejson==3.8.2
six==1.10.0
stripe==1.32.2
virtualenv==13.0.1
